add codingFooter component
